backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
openai-whisper>=20230314
pydub>=0.25.1
transformers>=4.30.0
torch>=2.0.0
torchaudio==2.2.0
vaderSentiment>=3.3.2
sentencepiece>=0.1.99 